feat(web/applications): add aria attributes to upload template
