+++ apps/web/src/client/components/file-uploader/file-uploader.module.js
@@ -1,0 +1,9 @@
+import registerUIEvents from './_ui.js';
+
+const allUploadForms = document.querySelectorAll('.pins-file-upload');
+
+const initFileUploaderModule = () => {
+	registerUIEvents(allUploadForms[0]);
+};
+
+export default initFileUploaderModule;
