+++ apps/web/src/server/views/app/components/file-uploader.component.njk
@@ -1,0 +1,53 @@
+{% from "govuk/components/inset-text/macro.njk" import govukInsetText %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+
+{% macro fileUploader(params) %}
+
+	<div class="govuk-grid-row">
+		{% if params.formTitle %}
+			<h2 class="govuk-heading-m govuk-grid-column-full">{{ params.formTitle }}</h2>
+		{% endif %}
+		<form class="pins-file-upload govuk-grid-column-two-thirds">
+			<div class="govuk-body colour--secondary pins-file-upload--instructions">
+				<label for="upload-file-{{ params.formId }}">Choose a single or multiple files to upload.</label>
+				<span>
+					Your file must be PDF, DOC, DOCX, PPT, PPTX, XLS, XLSX, JPG, JPEG, MPEG,
+					MP3, MP4, MOV, PNG, TIF, TIFF or ZIP.
+				</span>
+				<span>The total of your uploaded files must be smaller than 1 GB.</span>
+			</div>
+			<div class="display--flex">
+				<input class="display--none"
+					   id="upload-file-{{ params.formId }}"
+					   type="file" name="files" value="Choose file"
+					   aria-controls="file-list-{{ params.formId }}"
+					   aria-label="test A"
+					   multiple>
+				<button type='button' class="pins-file-upload--button govuk-button">
+					Choose file
+				</button>
+				<span role="status" aria-live="assertive" class='govuk-body pins-file-upload--counter'>
+					No file chosen
+				</span>
+			</div>
+			<div id="file-list-{{ params.formId }}">
+				<h3 class="display--sr-only" id="file-list-title-{{ params.formId }}">List of files to upload</h3>
+				<ul role="list" class="pins-file-upload--files-rows" aria-describedby="file-list-title-{{ params.formId }}" aria-live="polite">
+					<li role="listitem" class="pins-file-upload--file-row display--none">
+						<p class="govuk-heading-s" aria-details="File name"></p>
+						<button type="button" role="button" class="govuk-link pins-file-upload--remove" aria-details="Remove added file from list">Remove</button>
+					</li>
+				</ul>
+			</div>
+
+			{% if params.disclaimerText %}
+				{{ govukInsetText({
+					text: params.disclaimerText
+				}) }}
+			{% endif %}
+
+			{{ govukButton({ text: 'Save and continue' }) }}
+		</form>
+	</div>
+
+{% endmacro %}
