+++ apps/web/src/server/views/appeals/case-officer/appeal-documents.njk
@@ -3,7 +3,7 @@
 {% from "govuk/components/breadcrumbs/macro.njk" import govukBreadcrumbs %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "app/components/document-list.component.njk" import documentList %}
-{% from "app/components/file-uploader.njk" import fileUploader %}
+{% from "app/components/file-uploader_legacy.njk" import fileUploader %}
 
 {% set documentsForType = documents | filter({ Type: documentType }) %}
 
@@ -22,7 +22,7 @@
 				{ href: '/appeals-service/case-officer/appeals/' + appeal.AppealId, text: appeal.AppealReference },
         { text: pageTitle }
 			]
-		})	
+		})
 	}}
 {% endblock %}
