+++ apps/web/src/styles/4-generic/_generic.display.scss
@@ -8,3 +8,14 @@ $display-attrs: (
 	display: #{$val} !important;
   }
 }
+
+.display--sr-only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0,0,0,0);
+  border: 0;
+}
